+++ src/ui/svelte/apps-menu/index.ts
@@ -12,6 +12,7 @@ await loadTranslations();
 const widget = Widget.getCurrent();
 
 await widget.init({ hideOnFocusLoss: true });
+await widget.window.setFocusable(true);
 
 widget.onTrigger(async (args) => {
   const visible = await widget.window.isVisible();